Deconstruction of VFX
